- bbsengine6/session.py:   * add '**kwargs' to all functions   * session.start() now has a _work(), and the function can make a database connection if it was passed a pool (standard)   * added garbagecollect()

+++ py/src/bbsengine6/session.py
@@ -27,96 +27,139 @@ def build(rec):
 
     return session
 
-def start(args):
+def start(args, **kwargs):
     global currentsessionid
 
-    with database.connect(args) as conn:
-        garbagecollect(args)
+    def _work(args, conn):
+        global currentsessionid
+
+        garbagecollect(args, **kwargs)
+
         if currentsessionid is None: # and exists in the database
-            if args.debug is True:
-                io.echo("session.start.100: currentsessionid is None")
-            session = getmembersession(args)
+            io.echo("session.start.100: currentsessionid is None", level="debug")
+            session = getmembersession(args, **kwargs)
             if session is False:
                 io.echo("multiple sessions detected", level="error")
                 return False
             if session is None:
-                if args.debug is True:
-                    io.echo("session.start.120: creating new session")
-                session = buildsession(args)
+                io.echo("session.start.120: creating new session", level="debug")
+                session = buildsession(args, **kwargs)
                 if args.debug is True:
                     mogrify = True
                 else:
                     mogrify = False
-#                io.echo(f"bbsengine6.session.start.120: {session=}", level="debug")
-                currentsessionid = database.insert(args, "engine.__session", session, mogrify=mogrify)
+                    # io.echo(f"bbsengine6.session.start.120: {session=}", level="debug")
+                currentsessionid = database.insert(args, "engine.__session", session, mogrify=mogrify, **kwargs)
                 conn.commit()
                 return True
             else:
-                if args.debug is True:
-                    io.echo(f"bbsengine6.session.start.140: {session=}", level="debug")
+                io.echo(f"bbsengine6.session.start.140: {session=}", level="debug")
                 currentsessionid = session["id"]
         else:
-            if args.debug is True:
-                io.echo(f"bbsengine6.session.start.160: reading {currentsessionid=}", level="debug")
-            session = read(args, currentsessionid) # getmembersession(args, member.getcurrentid())
+            io.echo(f"bbsengine6.session.start.160: reading {currentsessionid=}", level="debug")
+            session = read(args, currentsessionid, **kwargs) # getmembersession(args, member.getcurrentid())
             if session is None:
-                if args.debug is True:
-                    io.echo(f"read of session returned None", level="debug")
+                io.echo(f"read of session returned None", level="debug")
                 session = buildsession(args)
-                database.insert(args, session, sessionid)
+                database.insert(args, session, sessionid, **kwargs)
                 conn.commit()
                 return True
             if type(session) is list and len(session) > 1:
-                io.echo(f"multiple sessions for member {member.moniker!r} detected", level="error")
+                io.echo(f"multiple sessions for member {member.moniker=} detected", level="error")
                 return False
 
             currentsessionid = session["id"]
 
+    conn = kwargs.get("conn", None)
+    if conn is None:
+        if pool is None:
+            return False
+        with database.connect(args, **kwargs) as conn:
+            return _work(args, conn)
+    else:
+        return _work(args, conn)
+
     return True
 
-def getmembersession(args, moniker=None):
+def getmembersession(args, moniker=None, **kwargs):
+    def _work(conn):
+        sql = "select * from engine.__session where moniker=%s"
+        dat = (moniker,)
+        with database.cursor(conn) as cur:
+            cur.execute(sql, dat)
+            if cur.rowcount == 0:
+                return None
+            elif cur.rowcount > 1:
+                io.echo(f"multiple sessions for member {moniker=} found", level="error")
+                return False
+            rec = cur.fetchone()
+            io.echo(f"getmembersession.100: {rec=}", level="debug")
+            return build(rec)
+
     if moniker is None:
-        moniker = member.getcurrentmoniker(args)
+        moniker = member.getcurrentmoniker(args, **kwargs)
         if moniker is None:
             io.echo("getmembersession.100: You do not exist! Go Away!", level="error")
             return None
 
 #    garbagecollect(args)
 
-    with database.connect(args) as conn:
-        with database.cursor(conn) as cur:
-            sql = "select * from engine.__session where moniker=%s"
-            dat = (moniker,)
-            cur.execute(sql, dat)
-            if cur.rowcount == 0:
-                return None
-            elif cur.rowcount > 1:
-                io.echo(f"multiple sessions for member {moniker=} found", level="warn")
-                return False
-            rec = cur.fetchone()
-            io.echo(f"getmembersession.100: {rec=}", level="debug")
-            return build(rec)
+    conn = kwargs.get("conn", None)
+    if conn is None:
+        pool = kwargs.get("pool", None)
+        if pool is None:
+            io.echo(f"getmembersession.140: {pool=}", level="error")
+            return False
+        with database.connect(args, pool=pool) as conn:
+            return _work(conn)
+    else:
+        return _work(conn)
 
-def updatelastactivity(args, sessionid):
-    import os
+def updatelastactivity(args, sessionid, **kwargs):
+    def _work(conn):
+        import os
 
-    with database.connect(args) as conn:
-        session = read(args, sessionid)
+        session = read(args, sessionid, **kwargs)
         if session is None:
             return False
 
-        if args.debug is True:
-            io.echo(f"{session=}", level="log")
+        io.echo(f"bbsengine.session.updatelastactivity.120: {session=}", level="debug")
 
         session["lastactivity"] = "now()" # datetime.ctime()
         session["expiry"] = datetime.now() + timedelta(minutes=15)
         session["useragent"] = os.environ["TERM"] if "TERM" in os.environ else "NEEDINFO"
-        write(args, session, sessionid)
+        write(args, session, sessionid, **kwargs)
         conn.commit()
         return True
 
-def read(args, sessionid=None):
+
+    conn = kwargs.get("conn", None)
+    if conn is None:
+        pool = kwargs.get("pool", None)
+        if pool is None:
+            io.echo(f"bbsengine.updatelastactivity.100: {pool=}", level="error")
+            return False
+        with database.connect(args, pool=pool) as conn:
+            return _work(conn)
+    else:
+        return _work(conn)
+
+
+def read(args, sessionid=None, **kwargs):
     global currentsessionid
+
+    def _work(cur):
+        sql = "select * from engine.session where id=%s"
+        dat = (sessionid,)
+        cur.execute(sql, dat)
+        if args.debug is True:
+            io.echo(f"bbsengine6.session.140: {cur.rowcount=}", level="debug")
+        if cur.rowcount == 0:
+            if args.debug is True:
+                io.echo("bbsengine6.session.read.120: returning None", level="debug")
+            return None
+        rec = cur.fetchone()
+        return build(rec)
 
     if sessionid is None:
         if currentsessionid is None:
@@ -129,51 +172,45 @@ def read(args, sessionid=None):
     if args.debug is True:
         io.echo(f"bbsengine6.session.read.100: {sessionid=}", level="debug")
 
-    sql = "select * from engine.session where id=%s"
-    dat = (sessionid,)
-    with database.connect(args) as conn:
-        with database.cursor(conn) as cur:
-            cur.execute(sql, dat)
-            if args.debug is True:
-                io.echo(f"bbsengine6.session.140: {cur.rowcount=}", level="debug")
-            if cur.rowcount == 0:
-                if args.debug is True:
-                    io.echo("bbsengine6.session.read.120: returning None", level="debug")
-                return None
-            rec = cur.fetchone()
-            return build(rec)
+    conn = kwargs.pop("conn", None)
+    with database.cursor(conn, **kwargs) as cur:
+        return _work(cur)
 
-def write(args, session, sessionid=None):
+def write(args, session, sessionid=None, **kwargs):
     global currentsessionid
+    io.echo(f"bbsengine.session.write.220: {kwargs=}", level="debug")
 
 #    garbagecollect(args)
+
+    if sessionid is None:
+        if currentsessionid is None:
+            return False
+        sessionid = currentsessionid
 
     session["dateupdated"] = "now()"
     session["lastactivity"] = "now()"
 
     _session = copy.copy(session)
 
-    with database.connect(args) as conn:
-        if sessionid is None:
-            if currentsessionid is None:
-                return False
-            sessionid = currentsessionid
+    conn = kwargs.get("conn", None)
+    if conn is None:
+        io.echo("bbsengine.session.write.200: {conn=}", level="error")
+        return False
 
-        if args.debug is True:
-            io.echo(f"bbsengine6.session.write.100: {session=}", level="debug")
+    if args.debug is True:
+        io.echo(f"bbsengine6.session.write.100: {session=}", level="debug")
 
-        if "data" in _session and type(_session["data"]) is dict:
-            _session["data"] = database.Jsonb(_session["data"])
+    if "data" in _session and type(_session["data"]) is dict:
+        _session["data"] = database.Jsonb(_session["data"])
 
-        mogrify = True if args.debug is True else False
-        database.update(args, "engine.__session", sessionid, _session, mogrify=mogrify)
-        conn.commit()
-
-def buildsession(args, sessionid=None, data={}):
+    mogrify = True if args.debug is True else False
+    database.update(args, "engine.__session", sessionid, _session, mogrify=mogrify, **kwargs)
+    conn.commit()
+def buildsession(args, sessionid=None, data={}, **kwargs):
     if sessionid is None:
         sessionid = str(uuid.uuid1())
 
-    moniker = member.getcurrentmoniker(args)
+    moniker = member.getcurrentmoniker(args, **kwargs)
     if moniker is None:
         io.echo("buildsession.100: You do not exist! Go Away!", level="error")
         return None
@@ -191,8 +228,8 @@ def buildsession(args, sessionid=None, data={}):
         io.echo(f"bbsengine6.session.buildsession.100: {session=}", level="debug")
     return session
 
-def get(args, name:str, default=None, memberid:int=None):
-    session = getmembersession(args, memberid)
+def get(args, name:str, default=None, memberid:int=None, **kwargs):
+    session = getmembersession(args, memberid, **kwargs)
     if args.debug is True:
         io.echo(f"bbsengine6.session.get.100: {session=}", level="debug")
 
@@ -205,7 +242,22 @@ def get(args, name:str, default=None, memberid:int=None):
 
     return default
 
-def set(args, name, value, sessionid=None, memberid=None, reset=False, mogrify=True):
+def set(args, name, value, sessionid=None, memberid=None, reset=False, mogrify=True, **kwargs):
+    def _work(cur):
+        data = {}
+        data[name] = value
+
+        if args.debug is True:
+            io.echo(f"bbsengine6.session.set.100: {data=}", level="debug")
+
+        if reset is False:
+            sql = "update engine.__session set data=data||%s where id=%s"
+        else:
+            sql = "update engine.__session set data=%s where id=%s"
+
+        dat = (data, sessionid)
+        cur.execute(sql, dat)
+
     if sessionid is None:
         if currentsessionid is None:
             return False
@@ -217,33 +269,20 @@ def set(args, name, value, sessionid=None, memberid=None, reset=False, mogrify=T
             io.echo("bbsengine6.session.set.100: You do not exist! Go Away!", level="error")
             return None
 
-    data = {}
-#    if type(value) is dict:
-#        data[name] = Json(value)
-#    else:
-    data[name] = value
-
-    if args.debug is True:
-        io.echo(f"bbsengine6.session.set.100: {data=}", level="debug")
-
-    if reset is False:
-        sql = "update engine.__session set data=data||%s where id=%s"
-    else:
-        sql = "update engine.__session set data=%s where id=%s"
-
-    dat = (data, sessionid)
-    with database.connect(args) as conn:
-        with database.cursor(conn) as cur:
-            cur.execute(sql, dat)
-            conn.commit()
+    conn = kwargs.get("conn", None)
+    with database.cursor(conn) as cur:
+        _work(cur)
+#        conn.commit()
     return value
 
-def garbagecollect(args):
-    io.echo("bbsengine6.session.garbagecollect.100: running", level="debug")
-    sql = "delete from engine.__session where expiry < now()"
-    dat = ()
-    with database.connect(args) as conn:
-        with database.cursor(conn) as cur:
-            cur.execute(sql)
+def garbagecollect(args, **kwargs):
+    conn = kwargs.pop("conn", None)
+    if conn is None:
+        io.echo(f"bbsengine.session.garbagecollect.100: {conn=}", level="error")
+        return False
+    with database.cursor(conn, **kwargs) as cur:
+        sql = "delete from engine.__session where expiry < now()"
+        dat = ()
+        cur.execute(sql)
         conn.commit()
-    return
+    return True
